+(AnalyseMagnetization,Kagome*)
+ can now analyse Kagome*, Chain* and Honeycomb*
+ the adaptation for other lattices seems to be pretty straightforward
+ the AnalyseParameter handles the polymerization strength for the
Chain*. It won't work properly for Kagome* and Honeycomb* but this is
normal. One would only need to do a few changes to set things correctly
+ AnalyseMagnetization works for Kagome* and may also work for Chain*
and Honeycomb*.
+ Honeycomb0pp is running for SU3. Normally the results are close or
equal to Miklos's results for the 0pp part but it seems to be
problematic for the 000 part. One should do more tests with SU4.