reorganization of CreateSystem and Fermionic
+ the dependencies have been modified so MonteCarlo and CreateSystem are
independant. The connection between the two parts are the Fermionic and
Bosonic filesi, as it should be.
+ GenericSystem <- System1D, System2D
+ System1D <- Chain
+ System2D <- Square, Triangle, Honeycomb, Kagome
+ System1D and 2D should be able to select the correct eigenvectors with
Px = Py = 0 when the energy eigenvalues are degenerated
+ Fermionic doesn't contain H_ (=T_) anymore, now this is handled by
SystemdD. Fermionic only diagonalize H_ and check if it's degenerate. If
it is degenerate, a new GenericSystem<complex> will be created and the
SystemdD class will perform a full diagonalization (CSCO : complete set
of comuting observable)