+(System1D, Chain*, System2D)
+ System1D seems to finally solve the degeracy at the fermi level. It
selects the eigenvector that is a combination of k_x + (-k_x). This
allows the limit delta->0 to be correct.
+ For now, only ChainPolymerized works properly. ChainFermi should be
rewritten to get rid of the ChainFermi<complex> that would not be
useful anymore if it is impossible to find a way to correct the
artifitial di/tri-merization created. One way to solve this problem
might be to add a phase to the previously selected eigenvector.